EP round radius update
Using a fixed round radius ratio of 0.25 with maximum 0.25mm radius does
not make sense for the exposed pad as it is created without positive
fillet and therefore such a large radius would interfere with the
"lead".
Our research showed that using the same radius as the normal pads is a
much better option. There also has been a parameter added to directly
control the round radius from the size definition.